completed the run
i managed to complete the run. Now the script is able to take input of only IPV6 which has been given by ifconfig which i am attaching in the readme. Line 41 uses "id=3". i selected this because i had previously 2 other ip addresses in the forwarding page and as such, if you have only one, set yours to that and you should be good to go